Remove CHECKSUM files in archives.
I don't think CHECKSUM file is necessary when we've already had
CHECKSUM of the whole archive. The CHECKSUM file just makes the
process a lot more complicated to create and update the archive's
contents. Furthermore, the CHECKSUM files will overwrite each other
when the archives are extracted to a same directory.